<unistd.h>: Fix profil(2) prototype.
 First off the vm_offset_t is somewhat bogus in this context, same could
 be said about size_t variant in sys/sysproto.h.  Just use plain u_long
 type like it is already used in "struct uprof" and fix forth argument
 to take u_int.  In public headers prefer to use generic types.

+++ sys/sys/sysproto.h
@@ -301,7 +301,7 @@ struct	profil_args {
 #endif
 	caddr_t	samples;	char samples_[PAD_(caddr_t)];
 	size_t	size;	char size_[PAD_(size_t)];
-	size_t	offset;	char offset_[PAD_(size_t)];
+	u_long	offset;	char offset_[PAD_(u_long)];
 	u_int	scale;	char scale_[PAD_(u_int)];
 };
 struct	ktrace_args {
